… generated/api/integrations/splunk_service_vtproto.pb.go
… generated/api/v1/administration_events_service_vtproto.pb.go
… generated/api/v1/alert_service_vtproto.pb.go
… generated/api/v1/api_token_service_vtproto.pb.go
… generated/api/v1/audit_vtproto.pb.go
… generated/api/v1/auth_service_vtproto.pb.go
… generated/api/v1/authprovider_service_vtproto.pb.go
… generated/api/v1/cloud_source_service_vtproto.pb.go
… generated/api/v1/cluster_init_service_vtproto.pb.go
… generated/api/v1/cluster_service_vtproto.pb.go
… generated/api/v1/common_vtproto.pb.go
… generated/api/v1/compliance_management_service_vtproto.pb.go
… generated/api/v1/compliance_service_vtproto.pb.go
… generated/api/v1/config_service_vtproto.pb.go
… generated/api/v1/credential_expiry_service_vtproto.pb.go
… generated/api/v1/cve_service_vtproto.pb.go
… generated/api/v1/db_service_vtproto.pb.go
… generated/api/v1/debug_service_vtproto.pb.go
… generated/api/v1/declarative_config_health_service_vtproto.pb.go
… generated/api/v1/delegated_registry_config_service_vtproto.pb.go
… generated/api/v1/deployment_service_vtproto.pb.go
… generated/api/v1/detection_service_vtproto.pb.go
… generated/api/v1/discovered_cluster_service_vtproto.pb.go
… generated/api/v1/feature_flag_service_vtproto.pb.go
… generated/api/v1/group_service_vtproto.pb.go
… generated/api/v1/grpc_preference_service_vtproto.pb.go
… generated/api/v1/image_integration_service_vtproto.pb.go
… generated/api/v1/image_service_vtproto.pb.go
… generated/api/v1/integration_health_service_vtproto.pb.go
… generated/api/v1/metadata_service_vtproto.pb.go