### Description

With `unmarshal` we create a new slice that we copy pg data into. Then we unmarshal proto from this slice so we fully control that memory. This means we can reduce allocs by reusing memory that is already allocated for strings and bytes. This leads to significant allocs reduction (nearly 50% on getting many objects and 30% on a single alert)

> [unmarshal_unsafe generates a func (p *YourProto) UnmarshalVTUnsafe(data []byte) that behaves like UnmarshalVT, except it unsafely casts slices of data to bytes and string fields instead of copying them to newly allocated arrays, so that it performs less allocations. Data received from the wire has to be left untouched for the lifetime of the message. Otherwise, the message's bytes and string fields can be corrupted.](https://github.com/planetscale/vtprotobuf?tab=readme-ov-file#available-features:~:text=been%20newly%20allocated.-,unmarshal_unsafe,-generates%20a%20func)


```sql
benchstat -table "PGX"  -confidence 0.030  old.txt new.txt
PGX: 
                       │   old.txt   │               new.txt               │
                       │   sec/op    │   sec/op     vs base                │
Many/get_1_alerts-8      158.2µ ± 0%   157.2µ ± 1%        ~ (p=0.481 n=10)
Many/get_2_alerts-8      174.5µ ± 0%   172.6µ ± 0%        ~ (p=0.063 n=10)
Many/get_4_alerts-8      221.1µ ± 1%   211.0µ ± 0%   -4.57% (p=0.001 n=10)
Many/get_8_alerts-8      275.8µ ± 1%   252.9µ ± 0%   -8.31% (p=0.000 n=10)
Many/get_16_alerts-8     389.3µ ± 0%   344.9µ ± 1%  -11.40% (p=0.000 n=10)
Many/get_32_alerts-8     573.4µ ± 0%   502.5µ ± 0%  -12.36% (p=0.000 n=10)
Many/get_64_alerts-8     951.4µ ± 0%   865.9µ ± 0%   -8.98% (p=0.000 n=10)
Many/get_128_alerts-8    1.772m ± 0%   1.581m ± 0%  -10.78% (p=0.000 n=10)
Many/get_256_alerts-8    3.236m ± 0%   2.978m ± 0%   -7.99% (p=0.000 n=10)
Many/get_512_alerts-8    6.329m ± 1%   6.338m ± 1%        ~ (p=0.971 n=10)
Many/get_1024_alerts-8   12.56m ± 0%   10.82m ± 0%  -13.91% (p=0.000 n=10)
Many/get_2048_alerts-8   24.77m ± 0%   20.89m ± 1%  -15.66% (p=0.000 n=10)
Many/get_4096_alerts-8   50.13m ± 0%   40.77m ± 1%  -18.67% (p=0.000 n=10)
Many/get_8192_alerts-8   97.16m ± 0%   82.32m ± 1%  -15.28% (p=0.000 n=10)
geomean                  2.026m        1.835m        -9.43%

                       │    old.txt    │               new.txt                │
                       │     B/op      │     B/op      vs base                │
Many/get_1_alerts-8      11.981Ki ± 0%   9.685Ki ± 0%  -19.17% (p=0.000 n=10)
Many/get_2_alerts-8       19.65Ki ± 0%   15.05Ki ± 0%  -23.39% (p=0.000 n=10)
Many/get_4_alerts-8       34.99Ki ± 0%   25.80Ki ± 0%  -26.26% (p=0.000 n=10)
Many/get_8_alerts-8       66.43Ki ± 0%   48.05Ki ± 0%  -27.66% (p=0.000 n=10)
Many/get_16_alerts-8     129.61Ki ± 0%   92.86Ki ± 0%  -28.36% (p=0.000 n=10)
Many/get_32_alerts-8      257.2Ki ± 0%   183.6Ki ± 0%  -28.59% (p=0.000 n=10)
Many/get_64_alerts-8      511.3Ki ± 0%   364.3Ki ± 0%  -28.76% (p=0.000 n=10)
Many/get_128_alerts-8    1018.9Ki ± 0%   724.5Ki ± 0%  -28.89% (p=0.000 n=10)
Many/get_256_alerts-8     1.996Mi ± 0%   1.421Mi ± 0%  -28.83% (p=0.000 n=10)
Many/get_512_alerts-8     3.986Mi ± 0%   2.833Mi ± 0%  -28.94% (p=0.000 n=10)
Many/get_1024_alerts-8    8.002Mi ± 0%   5.695Mi ± 0%  -28.83% (p=0.000 n=10)
Many/get_2048_alerts-8    16.29Mi ± 0%   11.70Mi ± 0%  -28.20% (p=0.000 n=10)
Many/get_4096_alerts-8    32.46Mi ± 0%   23.28Mi ± 0%  -28.30% (p=0.000 n=10)
Many/get_8192_alerts-8    65.14Mi ± 0%   46.77Mi ± 0%  -28.21% (p=0.000 n=10)
geomean                   766.2Ki        556.6Ki       -27.36%

                       │   old.txt    │               new.txt               │
                       │  allocs/op   │  allocs/op   vs base                │
Many/get_1_alerts-8        185.0 ± 0%    136.0 ± 0%  -26.49% (p=0.000 n=10)
Many/get_2_alerts-8        285.0 ± 0%    187.0 ± 0%  -34.39% (p=0.000 n=10)
Many/get_4_alerts-8        484.0 ± 0%    288.0 ± 0%  -40.50% (p=0.000 n=10)
Many/get_8_alerts-8        883.0 ± 0%    491.0 ± 0%  -44.39% (p=0.000 n=10)
Many/get_16_alerts-8      1678.0 ± 0%    894.0 ± 0%  -46.72% (p=0.000 n=10)
Many/get_32_alerts-8      3.266k ± 0%   1.698k ± 0%  -48.01% (p=0.000 n=10)
Many/get_64_alerts-8      6.438k ± 0%   3.302k ± 0%  -48.71% (p=0.000 n=10)
Many/get_128_alerts-8    12.777k ± 0%   6.505k ± 0%  -49.09% (p=0.000 n=10)
Many/get_256_alerts-8     25.45k ± 0%   12.91k ± 0%  -49.28% (p=0.000 n=10)
Many/get_512_alerts-8     50.80k ± 0%   25.71k ± 0%  -49.39% (p=0.000 n=10)
Many/get_1024_alerts-8   101.50k ± 0%   51.32k ± 0%  -49.44% (p=0.000 n=10)
Many/get_2048_alerts-8    202.9k ± 0%   102.5k ± 0%  -49.46% (p=0.000 n=10)
Many/get_4096_alerts-8    405.7k ± 0%   205.0k ± 0%  -49.48% (p=0.000 n=10)
Many/get_8192_alerts-8    811.2k ± 0%   409.8k ± 0%  -49.49% (p=0.000 n=10)
geomean                   9.917k        5.384k       -45.71%
```

- [x] CHANGELOG update is not needed
- [x] Documentation is not needed
### Testing
- [x] inspected CI results
#### Automated testing
- [x] modified existing tests
- [x] contributed **no automated tests**
#### How I validated my change
CI